+++ insights-ui/src/app/tariff-reports/page.tsx
@@ -1,5 +1,6 @@
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
-import { fetchTariffReportsWithUpdatedAt, TariffIndustryDefinition } from '@/scripts/industry-tariff-reports/tariff-industries';
+import { fetchTariffReportsWithUpdatedAt } from '@/scripts/industry-tariff-reports/fetch-tariff-reports-with-updated-at';
+import { TariffIndustryDefinition } from '@/scripts/industry-tariff-reports/tariff-industries';
 import { BreadcrumbsOjbect } from '@dodao/web-core/components/core/breadcrumbs/BreadcrumbsWithChevrons';
 import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
 import { ChevronRight, FileText } from 'lucide-react';
