+++ insights-ui/src/scripts/industry-tariff-reports/tariff-industries.ts
@@ -1,7 +1,3 @@
-import { getS3KeyForIndustryTariffs } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
-import { TariffUpdatesForIndustry } from '@/scripts/industry-tariff-reports/tariff-types';
-import { getJsonFromS3, getJsonWithLastModifiedFromS3 } from '@/scripts/report-file-utils';
-
 export enum TariffIndustryId {
   plastic = 'plastic',
   automobiles = 'automobiles',
@@ -589,24 +585,6 @@ export function getTariffIndustryDefinitionById(industryId: TariffIndustryId): T
 
 export function fetchTariffReports(): TariffIndustryDefinition[] {
   return Object.values(TariffIndustryId).map((industryId) => getTariffIndustryDefinitionById(industryId));
-}
-
-const reportLastModifiedAtMap: Record<string, string> = {};
-export async function fetchTariffReportsWithUpdatedAt(): Promise<(TariffIndustryDefinition & { lastModified: string })[]> {
-  console.warn('fetchTariffReportsWithUpdatedAt is called which is quite expensive. Keep an eye.');
-  if (Object.values(reportLastModifiedAtMap).length === 0) {
-    const ids = Object.values(TariffIndustryId);
-    for (const industryId of ids) {
-      const key = getS3KeyForIndustryTariffs(industryId, 'tariff-updates.json');
-      const lastModified = (await getJsonWithLastModifiedFromS3<TariffUpdatesForIndustry>(key))?.lastModified || new Date().toISOString();
-      reportLastModifiedAtMap[industryId] = lastModified;
-    }
-  }
-
-  return Object.values(TariffIndustryId).map((industryId) => ({
-    ...getTariffIndustryDefinitionById(industryId),
-    lastModified: reportLastModifiedAtMap[industryId],
-  }));
 }
 
 export interface HeadingSubheadingCombination {
